// src/product/product-types.ts
import mongoose from "mongoose";

export interface Product {
  _id?: mongoose.Types.ObjectId;
  name: string;
  description: string;
  // todo: fix the price configuration type
  priceConfiguration: string;
  attributes: string;
  tenantId: string;
  categoryId: mongoose.Types.ObjectId;
  image: string;
  isPublish: boolean;
}

export interface Filter {
  tenantId?: string;
  categoryId?: mongoose.Types.ObjectId;
  isPublish?: boolean;
}

export interface PaginateQuery {
  page: number;
  limit: number;
}